backend/requirements.txt
fastapi
uvicorn
python-dotenv
torch
sentence-transformers
numpy
pandas
scikit-learn
requests
transformers